# Conflicts:
#	src/main/java/ru/skypro/homework/controller/UserController.java
#	src/main/java/ru/skypro/homework/service/UserService.java
#	src/main/java/ru/skypro/homework/service/impl/UserServiceImpl.java
#	src/main/java/ru/skypro/homework/service/mapper/impl/CommentMapperImpl.java

+++ src/main/java/ru/skypro/homework/service/impl/AdServiceImpl.java
@@ -1,11 +1,9 @@
 package ru.skypro.homework.service.impl;
 
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
-import ru.skypro.homework.dto.AdsDto;
-import ru.skypro.homework.dto.CreateAdsDto;
-import ru.skypro.homework.dto.FullAdsDto;
-import ru.skypro.homework.dto.ResponseWrapperAdsDto;
+import ru.skypro.homework.dto.*;
 import ru.skypro.homework.entity.Ad;
 import ru.skypro.homework.entity.Image;
 import ru.skypro.homework.entity.User;
@@ -21,23 +19,32 @@ import ru.skypro.homework.service.mapper.AdMapper;
 import java.io.IOException;
 import java.util.Collection;
 import java.util.UUID;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
+@RequiredArgsConstructor
 public class AdServiceImpl implements AdService {
 
+    private final Role role = Role.ADMIN;
     private final AdRepository adRepository;
     private final ImageRepository imageRepository;
     private final UserService userService;
     private final AdMapper adMapper;
 
-    public AdServiceImpl(AdRepository adRepository,
-                         ImageRepository imageRepository,
-                         UserService userService, AdMapper adMapper) {
-        this.adRepository = adRepository;
-        this.imageRepository = imageRepository;
-        this.userService = userService;
-        this.adMapper = adMapper;
+    /**
+     * Метод проверяет наличие доступа к объявлению по id
+     * @param id
+     * @see ru.skypro.homework.service.impl.AdServiceImpl
+     */
+    @Override
+    public boolean checkAccess(Integer id) {
+        Ad ad = adRepository.findById(id).orElseThrow(AdsNotFoundException::new);
+        Optional<User> user = userService.findAuthUser();
+        User notOptionalUser = user.get();
+        String currentPrincipalName = notOptionalUser.getUsername();
+        return ad.getAuthor().getUsername().equals(currentPrincipalName)
+                || notOptionalUser.getAuthorities().contains(role);
     }
 
     @Override
@@ -72,7 +79,7 @@ public class AdServiceImpl implements AdService {
 
     @Override
     public boolean removeAdDto(Integer id) {
-        if(adRepository.existsById(id)) {
+        if (checkAccess(id) && adRepository.existsById(id)) {
             adRepository.deleteById(id);
             return true;
         }
@@ -82,9 +89,12 @@ public class AdServiceImpl implements AdService {
     @Override
     public AdsDto updateAdDto(Integer id, CreateAdsDto createAdsDto) {
         Ad ad = adRepository.findById(id).orElseThrow(AdsNotFoundException::new);
-        ad.setDescription(createAdsDto.getDescription());
-        ad.setPrice(createAdsDto.getPrice());
-        ad.setTitle(createAdsDto.getTitle());
+        if (checkAccess(id)) {
+            ad.setDescription(createAdsDto.getDescription());
+            ad.setPrice(createAdsDto.getPrice());
+            ad.setTitle(createAdsDto.getTitle());
+            return adMapper.mapAdToAdDto(adRepository.save(ad));
+        }
         return adMapper.mapAdToAdDto(adRepository.save(ad));
     }
